parse_texte: improve article title detection
If the article title is in a table with more than
one row or column, it means it's not an article title
but it's part of an article.
In the future, the table itself should be parsed.

ex: https://www.senat.fr/leg/pjl11-203.html ("Article 135 de la loi n° 2008-1..")