Turn the repository into a Python package
There is a lot left to discussion, I only made `parse-texte` available for now given that it is my need and I am not sure which other parts are useful as scripts.

The `install_requires` requirements are hard coded for now and match those related to the `parse_texte` script but that can be improved, not sure if we want to let the option to install only parts of the package and so on.

Fix #62

CI will hopefully/probably complain :-)