update_procedure: simplify, document and add test (#95)
This commit started with the issue #95:
update_procedure was adding a debats_order to a promulgation step

First, I've added a test to make sure we don't have a regression.

Then, the bug was fixed mainly by using "in_discussion" as the current
step detector.

While doing that, I simplified/re-organized the code.