prepare_articles: add the case where the depot text URL doesn't match the Senate ID
fixed this one: https://www.senat.fr/dossier-legislatif/pjl2014-211.html

It crashed on compute_stats.
Why ? Because it couldn't not find the first step.
Why ? Because the debats_order was set to None for the first step.
Why ? Because prepare_articles didn't find the correct first step.
Why ? Because while finding the multi-depots it didn't find the first step.
Why ? Because it's hard-coded that the first text must be published under
      the same ID as the Senate ID.

So I added a better check to know if it's the first depot step.

Also extracted the re-ordering in a function.